add json ignore, refactor model structure
