+++ CameraServerTests/Mocks/MockedCameraRepository.cs
@@ -1,4 +1,4 @@
-﻿using CameraServer.Helpers;
+﻿using CameraServer.Models;
 using CameraServer.Repositories;
 using Sakur.WebApiUtilities.Models;
